Another Information About the Movie
Set in *Madrid*, a mysterious man known as the *"professor"* recruits a group of eight people, who choose city names as their aliases, to carry out an ambitious plan that involves entering the *Royal Mint of Spain*, and escaping with *€984 million*. After taking 67 people hostage inside the Mint, the team plans to remain inside for 11 days to print the money as they deal with elite police forces. In the events following the initial heist, the group's members are forced out of hiding and prepare for a second heist, with some additional members, this time aiming to escape with gold from the *Bank of Spain*, as they again deal with hostages and police forces.